beta_tr can now be in steps of 0.1 degrees

+++ src/nupyprop/scripts/run.py
@@ -144,7 +144,7 @@ def main():
     args = parser.parse_args()
 
     energies = np.power(10, np.fromstring(args.energy_val, dtype=float, sep=","))
-    angles = np.fromstring(args.angle_val, dtype=int, sep=",")
+    angles = np.fromstring(args.angle_val, dtype=float, sep=",")
 
     idepth = int(args.idepth_val)
     ch_lepton = str(args.lepton_id)
